pyproject.toml
[tool.poetry]
name = "pastebin"
version = "0.1.0"
description = "Pastebin pet-project"
authors = ["Denys <dpashnev175@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
redis = "^5.0.6"
loguru = "^0.7.2"
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.3.3"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
setuptools = "^70.1.1"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
ipython = "^8.25.0"
databases = "^0.9.0"
aiosqlite = "^0.20.0"
mypy = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
